Fixed pickle issues
Smac3 requires pickling of the multi processed function, which includes the dataloader. As pickling of lmdb environments and lambda functions is not possible, I replaced all lambda function based transforms with the actual transforms and moved the lmdb file loading into the function instead of in the initialization (which I think caused the lmdb env to be loaded early, forcing the pickle process to pickle the environment instead of the provesses creating their own environment.